Refactor to use Redux Ducks.
