+++ src/store/configureStore.prod.js
@@ -1,7 +1,7 @@
 import { createStore, applyMiddleware } from 'redux';
 import thunkMiddleware from 'redux-thunk';
 import { apiMiddleware } from 'redux-api-middleware';
-import rootReducer from 'reducers/index';
+import rootReducer from 'redux/modules/index';
 
 // Middleware you want to use in production:
 const enhancer = applyMiddleware(thunkMiddleware, apiMiddleware);
